requirements.txt
streamlit
requests
beautifulsoup4
pandas
python-dotenv
groq